requirements.txt
# see https://github.com/googleapis/python-bigquery/issues/206#issuecomment-1772608208
google-cloud-bigquery==3.27.0
google-cloud-bigquery-storage==2.27.0
db-dtypes==1.3.1
pandas==2.2.3
